+++ elfo-core/src/runtime.rs
@@ -3,6 +3,8 @@ use std::sync::Arc;
 use tokio::runtime::Handle;
 
 use crate::actor::ActorMeta;
+#[cfg(feature = "unstable-stuck-detection")]
+use crate::stuck_detection::StuckDetector;
 
 pub(crate) trait RuntimeFilter: Fn(&ActorMeta) -> bool + Send + Sync + 'static {}
 impl<F: Fn(&ActorMeta) -> bool + Send + Sync + 'static> RuntimeFilter for F {}
@@ -12,6 +14,8 @@ impl<F: Fn(&ActorMeta) -> bool + Send + Sync + 'static> RuntimeFilter for F {}
 #[derive(Default, Clone)]
 pub(crate) struct RuntimeManager {
     dedicated: Vec<(Arc<dyn RuntimeFilter>, Handle)>,
+    #[cfg(feature = "unstable-stuck-detection")]
+    stuck_detector: StuckDetector,
 }
 
 impl RuntimeManager {
@@ -27,5 +31,10 @@ impl RuntimeManager {
         }
 
         tokio::runtime::Handle::current()
+    }
+
+    #[cfg(feature = "unstable-stuck-detection")]
+    pub(crate) fn stuck_detector(&self) -> StuckDetector {
+        self.stuck_detector.clone()
     }
 }