+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- stuck_detection: a basic detector under the `unstable-stuck-detection` feature.
+
 ### Fixed
 - pinger: avoid getting stuck during reconfiguration.